% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahazwcal.R
\name{ahaz.wcal}
\alias{ahaz.wcal}
\title{Fit Additive Hazards Model from Nested Case-control Samples (NCC)}
\usage{
ahaz.wcal(
  outcome,
  etime,
  tau,
  ind.ph2,
  m,
  X,
  Z,
  incl.prob = NULL,
  nrisk = NULL,
  A = NULL,
  eta = NULL
)
}
\arguments{
\item{outcome}{binary outcome status (1 = case; 0 = non-case)}

\item{etime}{time-to-event (either case or censored, whichever comes first)}

\item{tau}{projection time for pure risk or beta coefficient}

\item{ind.ph2}{a vector of binary status whether included in NCC (phase 2) or not}

\item{m}{number of controls per case}

\item{X}{a matrix of model covariates with time-varying effects (Beta); one or more columns may include NA if missing by NCC}

\item{Z}{a matrix of model covariates with time-invariant effects (gamma); one or more columns may include NA if missing by NCC}

\item{incl.prob}{probability of being included in NCC; e.g., this can be computed from KMprob() of 'multipleNCC'; if NULL, it is internally computed based on outcome, etime, and m}

\item{nrisk}{number of at-risk subjects at each 'etime'; if NULL, it is internally computed based on outcome, etime, and m}

\item{A}{a matrix of auxiliary statistics; e.g., cbind(1, AUX) where AUX is an object of aux.ahaz(); if A = NULL, it returns estimation results with design weights i.e., without weight calibration}

\item{eta}{a vector of estimates of weight calibration parameters obtained from the attributes of a rakecal() object; if A is NULL, eta must also be NULL}
}
\value{
A list of beta-gamma estimates and their variance estimates with influence functions
}
\description{
This function fits the additive hazards model with time-varying coefficients (beta) and time-invariant coefficients (gamma) from nested case-control study with or without weight calibration.
Weight calibration parameters and auxiliary statistics are required to fit the model with weight calibration. Otherwise, it will provide the result without weight calibration.
}
\examples{
#### STEP1: Predict Missing Covariates ####
fit_Z2 = glm(Z2 ~ X1 + X2 + Z1 + U, family = binomial(link = "logit"), 
             data = sample_data, subset = (ind.ph2 == 1), weights = 1/incl.prob)

sample_data.tilde = sample_data
sample_data.tilde$Z2 = predict(fit_Z2, type = "response", newdata = sample_data)


#### STEP2: Obtain Auxiliary Statistics (Influence Functions) ####
aux = with(sample_data.tilde, aux.ahaz(outcome = ind.fail, etime = eventime, tau = 8, X = cbind(X1, X2), Z = cbind(Z1, Z2)))

aux1 = cbind(1, aux[,c("gamma1","gamma2")]) # Auxiliary 1: Influence functions of gamma estimates 
aux2 = cbind(1, aux) # Auxiliary 2: Influence functions of Beta and gamma estimates


#### STEP3: Calibrate Weights ####
cal1 = with(sample_data.tilde, rakecal(A = aux1[ind.ph2,], w0 = 1/incl.prob[ind.ph2], 
                                       total = colSums(aux1), eta = T, EPS = 1e-12))
cal2 = with(sample_data.tilde, rakecal(A = aux2[ind.ph2,], w0 = 1/incl.prob[ind.ph2], 
                                       total = colSums(aux2), eta = T, EPS = 1e-12))


#### STEP4: Fit Additive Hazards Model ####
# with design weights; no calibration 
fit_ahaz_w = with(sample_data, ahaz.wcal(outcome = ind.fail, etime = eventime, tau = 8, 
                                 X = cbind(X1, X2), Z = cbind(Z1, Z2),
                                 ind.ph2 = ind.ph2, incl.prob = incl.prob, m = 1, nrisk = nrisk,
                                 A = NULL, eta = NULL))
                                 
# with calibrated weights with aux1
fit_ahaz_wcal_g = with(sample_data, ahaz.wcal(outcome = ind.fail, etime = eventime, tau = 8, 
                                      X = cbind(X1, X2), Z = cbind(Z1, Z2),
                                      ind.ph2 = ind.ph2, incl.prob = incl.prob, m = 1, nrisk = nrisk,
                                      A = aux1, eta = attr(cal1, "eta")))
                                      
# with calibrated weights with aux2
fit_ahaz_wcal_Bg = with(sample_data, ahaz.wcal(outcome = ind.fail, etime = eventime, tau = 8, 
                                       X = cbind(X1, X2), Z = cbind(Z1, Z2),
                                       ind.ph2 = ind.ph2, incl.prob = incl.prob, m = 1, nrisk = nrisk,
                                       A = aux2, eta = attr(cal2, "eta")))
                                       
#### Result: Estimates of Beta and gamma Coefficients ####
cbind(Estimate = fit_ahaz_w$est, StandardError = fit_ahaz_w$se)
cbind(Estimate = fit_ahaz_wcal_g$est, StandardError = fit_ahaz_wcal_g$se)
cbind(Estimate = fit_ahaz_wcal_Bg$est, StandardError = fit_ahaz_wcal_Bg$se)


}
\seealso{
\code{\link{sample_data}}, \code{\link{aux.ahaz}}, \code{\link{rakecal}}
}
\keyword{ahazcal}
