% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxahaz.R
\name{aux.ahaz}
\alias{aux.ahaz}
\title{Auxiliary Statistics of Pure Risk Estimations from Additive Hazards Model}
\usage{
aux.ahaz(outcome, etime, tau, X, Z)
}
\arguments{
\item{outcome}{binary outcome status (1 = case; 0 = non-case)}

\item{etime}{time to event (either case or censored, whichever comes first)}

\item{tau}{risk projection time}

\item{X}{a matrix of observed or predicted model covariates with time-varying effects (Beta)}

\item{Z}{a matrix of observed or predicted model covariates with time-invariant effects (gamma)}
}
\value{
a matrix of auxiliary statistics of pure risk estimates (cumulated influence functions of beta and influence functions of gamma) to calibrate design weights
}
\description{
This function provides influence functions of additive hazards model coefficients: beta are time-varying coefficients of X; and gamma are time-invariant coefficients of Z. 
The resulting values with observed/predicted covariates in a full cohort can be used as auxiliary statistics of pure risk estimates for weight calibration.
All arguments for this function except tau must be complete with the same length of vectors.
}
\examples{
#### STEP1: Predict Missing Covariates ####
fit_Z2 = glm(Z2 ~ X1 + X2 + Z1 + U, family = binomial(link = "logit"), 
             data = sample_data, subset = (ind.ph2 == 1), weights = 1/incl.prob)

sample_data.tilde = sample_data
sample_data.tilde$Z2 = predict(fit_Z2, type = "response", newdata = sample_data)


#### STEP2: Obtain Auxiliary Statistics (Influence Functions) ####
aux = with(sample_data.tilde, aux.ahaz(outcome = ind.fail, etime = eventime, tau = 8, 
                                       X = cbind(X1, X2), Z = cbind(Z1, Z2)))

aux1 = cbind(1, aux[,c("gamma1","gamma2")]) # Auxiliary 1: Influence functions of gamma estimates 
aux2 = cbind(1, aux) # Auxiliary 2: Influence functions of Beta and gamma estimates

}
\seealso{
\code{\link{sample_data}}
}
\keyword{ahazcal}
