% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{jointVP.ncc}
\alias{jointVP.ncc}
\title{Joint Probabilities of Nested Case-control Samples}
\usage{
jointVP.ncc(outcome, etime, tau, m, incl.prob = NULL, nrisk = NULL)
}
\arguments{
\item{outcome}{binary outcome status (1 = case; 0 = non-case)}

\item{etime}{time to event (either case or censored, whichever comes first)}

\item{tau}{risk projection time}

\item{m}{number of controls per case}

\item{incl.prob}{probability of being included in NCC; e.g., this can be computed from KMprob() of 'multipleNCC'; if NULL, it is internally computed based on outcome, etime, and m}

\item{nrisk}{= number of at-risk subjects at each 'etime'; if NULL, it is internally computed based on outcome, etime, and m}
}
\description{
Internal use to compute joint probabilities of nested case-control samples. The source code is similar to 'multipleNCC' package.
}
\keyword{ahazcal}
