% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sqdg}
\alias{sqdg}
\title{Extract Standard Deviation from A Variance-covariance Matrix}
\usage{
sqdg(x)
}
\arguments{
\item{x}{A n x n variance-covariance matrix}
}
\description{
Internal use to compute the square roots of diagonals of a given variance-covariance matrix.
}
\keyword{ahazcal}
