% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wcrossprod}
\alias{wcrossprod}
\title{Weighted Matrix Crossproduct}
\usage{
wcrossprod(x, w, y = NULL)
}
\arguments{
\item{x}{n x p matrix}

\item{w}{n vector of weights}

\item{y}{n x q matrix; if NULL, y = x}
}
\description{
Internal use to compute the weighted crossproduct of one or two matrices. (x'wy)
}
\keyword{ahazcal}
